+++ eca-chat-expandable.el
@@ -289,7 +289,7 @@ NESTED-PROPS is a plist with :parent-id and :label-indent for nested blocks."
                                                                                                             (make-string (length eca-chat-expandable-block-open-symbol) ?\s)))))
                                   'keymap (let ((km (make-sparse-keymap)))
                                             (define-key km (kbd "<mouse-1>") (lambda () (interactive) (eca-chat--expandable-content-toggle id)))
-                                            (define-key km (kbd "<tab>") (lambda () (interactive) (eca-chat--expandable-content-toggle id)))
+                                            (define-key km (kbd "TAB") (lambda () (interactive) (eca-chat--expandable-content-toggle id)))
                                             (define-key km (kbd "RET") (lambda () (interactive) (eca-chat--expandable-content-toggle id)))
                                             km)
                                   'help-echo "mouse-1 / tab / RET: expand/collapse"))